+++ System/Hardware/Arduino/SamplePrograms/Morse.hs
@@ -1,115 +1,78 @@
-module Morse where
+-------------------------------------------------------------------------------
+-- |
+-- Module      :  System.Hardware.Arduino.SamplePrograms.Morse
+-- Copyright   :  (c) Antoine R. Dumont, Levent Erkok
+-- License     :  BSD3
+-- Maintainer  :  erkokl@gmail.com
+-- Stability   :  experimental
+--
+-- Morse code blinker. Original by Antoine R. Dumont, modified to simplify
+-- and fit into the existing examples structure.
+-------------------------------------------------------------------------------
+module System.Hardware.Arduino.SamplePrograms.Morse where
 
-import qualified Data.Map as Map
+import Control.Monad       (forever)
+import Control.Monad.Trans (liftIO)
+import Data.Char           (toUpper)
+import Data.List           (intercalate)
+import Data.Maybe          (fromMaybe)
+
 import System.Hardware.Arduino
-import System.Environment
-import Data.Char (toLower)
 
--- unit of time (ms)
-unitOfTime :: Int
-unitOfTime = 300
+-- | A dit or a dah is all we need for Morse:
+-- A @dit@ is a dot; and a @dah@ is a dash in the Morsian world.
+-- We use 'LBreak' and 'WBreak' to indicate a letter and a word break
+-- so we can insert some delay between letters and words as we
+-- transmit.
+data Morse = Dit | Dah | LBreak | WBreak
+           deriving Show
 
--- short pulse
-dit :: Int
-dit = 1 * unitOfTime
+-- | Morse code dictionary
+dict :: [(Char, [Morse])]
+dict = map encode m
+  where encode (k, s) = (k, map (\c -> if c == '.' then Dit else Dah) s)
+        m = [ ('A', ".-"   ), ('B', "-..." ), ('C', "-.-." ), ('D', "-.."  ), ('E', "."    )
+            , ('F', "..-." ), ('G', "--."  ), ('H', "...." ), ('I', ".."   ), ('J', ".---" )
+            , ('K', "-.-"  ), ('L', ".-.." ), ('M', "--"   ), ('N', "-."   ), ('O', "---"  )
+            , ('P', ".--." ), ('Q', "--.-" ), ('R', ".-."  ), ('S', "..."  ), ('T', "-"    )
+            , ('U', "..-"  ), ('V', "...-" ), ('W', ".--"  ), ('X', "-..-" ), ('Y', "-.--" )
+            , ('Z', "--.." ), ('0', "-----"), ('1', ".----"), ('2', "..---"), ('3', "...--")
+            , ('4', "....-"), ('5', "....."), ('6', "-...."), ('7', "--..."), ('8', "---..")
+            , ('9', "----."), ('+', ".-.-."), ('/', "-..-."), ('=', "-...-")
+            ]
 
--- long pulse
-dahs :: Int
-dahs = 3 * unitOfTime
+-- | Given a sentence, decode it. We simply drop any letters that we
+-- do not have a mapping for.
+decode :: String -> [Morse]
+decode = intercalate [WBreak] . map (intercalate [LBreak] . map cvt) . words
+ where cvt c = fromMaybe [] $ toUpper c `lookup` dict
 
-pauseBetweenLetters :: Int
-pauseBetweenLetters = 3 * unitOfTime
+-- | Given a morsified sentence, compute the delay times. A 'Left' value means
+-- turn the led on that long, a 'Right' value means turn it off that long.
+morsify :: [Morse] -> [Either Int Int]
+morsify = map t
+  where unit     = 300
+        t Dit    = Left  $ 1 * unit
+        t Dah    = Left  $ 3 * unit
+        t LBreak = Right $ 3 * unit
+        t WBreak = Right $ 7 * unit
 
-pauseBetweenWords :: Int
-pauseBetweenWords = 7 * unitOfTime
+-- | Finally, turn a full sentence into a sequence of blink on/off codes
+transmit :: Pin -> String -> Arduino ()
+transmit p = sequence_ . concatMap code . morsify . decode
+  where code (Left i)  = [digitalWrite p True,  delay i, digitalWrite p False, delay i]
+        code (Right i) = [digitalWrite p False, delay i]
 
--- http://www.colorpilot.com/morse.html
-
--- map representation of the morse
-morseMap :: Map.Map Char [Integer]
-morseMap = Map.fromList [('a', [0,1]),
-                         ('b', [1,0,0,0]),
-                         ('c', [1,0,1,0]),
-                         ('d', [1,0,0]),
-                         ('e', [0]),
-                         ('f', [0,0,1,0]),
-                         ('g', [1,1,0]),
-                         ('h', [0,0,0,0]),
-                         ('i', [0,0]),
-                         ('j', [0,1,1,1]),
-                         ('k', [1,0,1]),
-                         ('l', [0,1,0,0]),
-                         ('m', [1,1]),
-                         ('n', [1,0]),
-                         ('o', [1,1,1]),
-                         ('p', [0,1,1,0]),
-                         ('q', [1,1,0,1]),
-                         ('r', [0,1,0]),
-                         ('s', [0,0,0]),
-                         ('t', [1]),
-                         ('u', [0,0,1]),
-                         ('v', [0,0,0,1]),
-                         ('w', [0,1,1]),
-                         ('x', [1,0,0,1]),
-                         ('y', [1,0,1,1]),
-                         ('z', [1,1,0,0])]
-
-type Bit = Integer
-type Word = String
-type Sentence = [Word]
-
-type MorseSentence = [[[Bit]]]
-type MorseWord = [[Bit]]
-type MorseLetter = [Bit]
-
-wordToMorse :: Word -> MorseWord
-wordToMorse =
-  map morseWord
-  where morseWord w = case (Map.lookup (toLower w) morseMap) of
-          Just v -> v
-          _      -> []
-
-sentenceToMorse :: Sentence -> MorseSentence
-sentenceToMorse = map wordToMorse
-
--- arduino part
-
--- prepare the led
-prepareLed :: Pin -> Arduino ()
-prepareLed = flip setPinMode OUTPUT
-
--- make the led blink for a given time
-blink :: Pin -> Int -> Arduino ()
-blink l pauseTime = do digitalWrite l True
-                       delay pauseTime
-                       digitalWrite l False
-                       delay pauseTime
-
-speakLetter :: Pin -> MorseLetter -> Arduino ()
-speakLetter l bs = mapM_ (blink l . computePauseTime) bs
-                   where computePauseTime 1 = dahs
-                         computePauseTime 0 = dit
-
-speakWord :: Pin -> MorseWord -> Arduino ()
-speakWord l bbs = mapM_ doSpeakLetter bbs
-                  where doSpeakLetter bs = do speakLetter l bs
-                                              delay pauseBetweenLetters
-
-speakSentence :: Pin -> MorseSentence -> Arduino ()
-speakSentence l bbbs = mapM_ doSpeakWord bbbs
-                       where doSpeakWord bbs = do speakWord l bbs
-                                                  delay pauseBetweenWords
-
--- Make the led blink in morse the given sentence
-speakMorseWithDevice :: FilePath -> String -> IO ()
-speakMorseWithDevice device sentence = withArduino False device $ do
-                         let led = digital 13
-                         prepareLed led
-                         speakSentence led $ (sentenceToMorse . words) sentence
-                         return ()
-
--- run from the cli
-main :: IO ()
-main =
-  do (devicePath:sentence:_) <- getArgs
-     speakMorseWithDevice devicePath sentence
+-- | A simple demo driver. To run this example, you only need the Arduino connected to your
+-- computer, no other hardware is needed. We use the internal led on pin 13. Of course,
+-- you can attach a led to pin 13 as well, for artistic effect.
+--
+--  <<http://github.com/LeventErkok/hArduino/raw/master/System/Hardware/Arduino/SamplePrograms/Schematics/Blink.png>>
+morseDemo :: IO ()
+morseDemo = withArduino False "/dev/cu.usbmodemfd131" $ do
+                setPinMode led OUTPUT
+                forever send
+ where  led  = digital 13
+        send = do liftIO $ putStr "Message? "
+                  m <- liftIO getLine
+                  transmit led m